Various unit test improvements
